backend/requirements.txt
PyYAML>=6.0
Pydantic[email]>=2.0
Jinja2>=3.1
WeasyPrint>=60.0
Pillow>=10.0
